JPERF-903 Open NotificationPopUps class to make it accessible from other JPT related libraries
To fix the flakiness of edit sprint action defined in jira-software-actions module I needed a way to close AUI flags. The mechanism for that was already implemented in this module however, NotificationPopUps was defined as an internal class. To make it possible to use this class in jira-software-action and avoid code duplication I removed that modifier.