JPERF-1043: Coalesce JQL actions together
The new action is better, because:
* it supports `IssueNavResultsView`
* doesn't copy-paste so much code
* doesn't call `IssueKeyMemory.remember` twice
* doesn't rely on smelly `JqlMemory.recallByTag` API
* distinguishes counting "no issues found" and "unknown issues found"

I call `recallByTag` smelly, because:
* there's no such thing as "JQL tags" for the users
* it's way too easy to think `Predicate<String>` filters JQLs, not tags
* it's bound too tightly to `AdaptiveJqlMemory`

But I'm not deprecating it here, because I'd need to produce a
replacement, which basically means decoupling `AdaptiveJqlMemory`
into "JQL baking" and "JQL filtering", but it's too much for this scope.

PS. Restore binary compatibility of constructors of:
* `SearchJqlAction`
* `SearchJqlChangelogAction`
* `SearchJqlSimpleAction`

Adding Kotlin default parameters doesn't preserve the method signatures.