JPERF-1043: Measure issue nav view switches
Measurement will tell us the error rate and latency
of both the view detection and view switching.

Observations will tell us how often the switch actually switched.
And if switching is consistent with the detected view before and after.

Also, it allows to observe alternative actions, which would only detect
the view, without trying to switch it.