JPERF-149: Stop expecing iframe on dashboard
Dashboards can have any number of gadgets, from zero to hundreds.
And gadgets don't need to use iframes at all.

The original commit message said:
> I'm using a hack to check if iframes on dashboard are loaded.
> This is intended to emulate browser metrics on dashboard.
> I did not use browser metrics as those are flakeing on my hardware

Back in the day we were aligning Action Metrics (from VU)
with Browser Metrics (from Jira). However, BM coverage was spotty,
changing and flaky. Moreover, even if we would harden BM, it would
suffer from a fundamental problem of coordinated ommission.
That is, when Jira frontend is not performing well, it is not sending
analytics well, so problems are systematically hidden.
Even if FE is working, backend performance and bugs can still
systematically skip the bad performance measurements, e.g. when the
threads are starved, GC overhead is blocking analytics/JSON processing
or writing to disk is blocked.

Therefore we need an external view of performance, without relying on
Jira systems. That's what VUs provide with Action Metrics.