JPERF-1435: Reduce JS overhead
For every resource (navigation, XHR, JS, CSS, image), it would send a
new HTTP request to parse headers.

The resource URI was not taken into account in the synthetic request,
so it could not infer original HTTP headers. Especially since the new
request can have different response headers than the original requests.
And even if they had the same, then it means we assume sticky session.
So we might read node id once just as well.

Supposedly, the synthetic request was being read from a cache,
but apparently it's still too much overhead.

We'll have to read the node id from `Server-Timing` header,
which is automatically captured by the browser
and it's the same way we obtain the response thread id.